requirements.txt
Flask
tensorflow
nump
Pillow
opencv-python
